Introduce Makefiles for building
With these in place everything builds with the Android NDK.  There is
also a build for host machines for the purpose of testing.  Needed to
bring in a uuid.h since the Android NDK lacks the functionality.  This
should be usable once built.  The merge strategies are incomplete, still
need the functionality of fixing things so the CRC remains the same.  The
intermediate merge strategy also needs to be completed.