Rearrange Makefiles, get test code operational on Android device
Make can be something of a beast to get it to do what you want it to.
Having broken the GPT test code into a separate file, it is now rather
easier to have the gpt test/display code run on an Android device.  I
kind of dislike having it outside gpt.c, but since this makes building
sane...