Appease clang
I'm pretty sure clang is right to complain about the GPT header being
marked as extendable and then being embedded in the middle of another
structure.  This is annoying since the extension is meant to represent
potential on-media data which is unlikely to ever be in the real
structure, but appease clang anyway.