Tweak replacegpt() slightly, 3 bugfixed in mergegpt()
I suspect in general both the old and new GPTs will have the same entry
counts and this fix will be irrelevant, but I could be surprised.  If the
number differs, this loop could go off into whacky memory locations.

Also needed to check for running out of empty entries to reuse.  This
should also be unlikely, but quite possible.  In such situation we don't
want to SEGV.