First plausible implementation of hybridgpt()
Been aiming for this goal for a while, trying to get this alternative
strategy working.  Much of this function is pretty similar to mergegpt(),
but there are some small differences.  Due to better knowledge of the
Android libraries the approach of this code seems slightly better.
Testing is needed before this is inflicted upon anyone.