Fix multiple files at once functionality
Oh gawd, that was stupid.  That was really stupid.  This is just one of
those bugs that you can't believe you did something so basically wrong.
Okay, multiple file mode works now and will report on all the files, not
one copy of the report for the first file for each file specified.